Add circle CI configuration
Ensure we run the tests on a machine other than mine. This will make
getting the project up and running simpler (because this will run that
process), but also give us confidence we're not introducing regressions.

This also removes the healthcheck, which is not supported on the version
of Docker that is in CircleCI